refactor : main activity 수정, nav default 변경, homescreen 추가
