--- app/src/main/java/com/example/sachosaeng/MainActivity.kt
@@ -1,45 +1,0 @@
-package com.example.sachosaeng
-
-import android.content.ContentValues.TAG
-import android.os.Bundle
-import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.sachosaeng.ui.theme.SachosaengTheme
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.common.util.Utility
-import com.kakao.sdk.user.UserApiClient
-import kotlin.jvm.functions.Function2
-
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            SachosaengTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    Button(
-                        onClick = { },
-                        modifier = Modifier.padding(16.dp)
-                    ) {
-                        Text("Button")
-                    }
-                }
-            }
-        }
-    }
-}
