+++ app/src/main/java/com/example/sachosaeng/feature/home/HomeScreen.kt
@@ -1,0 +1,9 @@
+package com.example.sachosaeng.feature.home
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeScreen() {
+    Text(text = "Home Screen")
+}
