+++ app/src/main/java/com/example/sachosaeng/main/MainViewModel.kt
@@ -1,0 +1,12 @@
+package com.example.sachosaeng.main;
+
+import android.content.ContentValues
+import android.util.Log
+import androidx.lifecycle.ViewModel
+import com.kakao.sdk.user.UserApiClient
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+
+class MainViewModel @Inject constructor() : ViewModel() {
+
+}
